change effect layer from byte[] to int[]

+++ src/main/java/com/astune/painter/item/EffectCreator.java
@@ -204,7 +204,7 @@ public class EffectCreator extends Item implements IPaintProvider {
         @Override
         public NativeImage createImage(CanvasFace face) {
             //System.out.println("glow image");
-            byte[] glowLayer = face.getEffectLayer("glow");
+            int[] glowLayer = face.getEffectLayer("glow");
             if (glowLayer == null) return null;
 
             PixelMatrix matrix = face.pixels();
@@ -222,7 +222,7 @@ public class EffectCreator extends Item implements IPaintProvider {
             try {
                 for (int y = 0; y < h; y++) {
                     for (int x = 0; x < w; x++) {
-                        if (glowLayer[y * w + x] == (byte)255){
+                        if (glowLayer[y * w + x] == 255){
                             int argb = matrix.getPixel(x, y);
                             int a = (argb >> 24) & 0xFF;
                             int r = (argb >> 16) & 0xFF;
